+++ fj-doc-mod-openpdf-ext/src/main/java/org/fugerit/java/doc/mod/openpdf/ext/PdfTypeHandler.java
@@ -10,6 +10,7 @@ import org.fugerit.java.doc.base.config.DocOutput;
 import org.fugerit.java.doc.base.config.DocTypeHandler;
 import org.fugerit.java.doc.base.config.DocTypeHandlerDefault;
 import org.fugerit.java.doc.base.model.DocBase;
+import org.fugerit.java.doc.mod.openpdf.ext.helpers.DocumentMetaHelper;
 import org.fugerit.java.doc.mod.openpdf.ext.helpers.OpenPDFConfigHelper;
 import org.fugerit.java.doc.mod.openpdf.ext.helpers.OpenPpfDocHandler;
 import org.w3c.dom.Element;
@@ -57,6 +58,7 @@ public class PdfTypeHandler extends DocTypeHandlerDefault {
 			pdfWriter = PdfWriter.getInstance( document, baos );
 			handler = new OpenPpfDocHandler(document, pdfWriter, totalPageCount );
 		}
+		DocumentMetaHelper.handleDocMeta( document, docBase );
 		handler.handleDoc( docBase );
 		baos.writeTo( outputStream );
 		baos.close();