+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+[fj-doc-mod-openrtf-ext] handling doc-title, doc-subject, doc-author, doc-language, doc-creator #196
+[fj-doc-mod-openpdf-ext] handling doc-title, doc-subject, doc-author, doc-language, doc-creator #196
+[fj-doc-mod-openpdf-ext] default creator set to 'OpenPDF over Fugerit Venus DOC' #196
 [fj-doc-mod-fop] default creator set to 'Apache FOP over Fugerit Venus DOC'
 
 ## [8.8.3] - 2024-09-11